Refactor newsletter management by restructuring resources
- Moved NewsletterResource to a new namespace: Firefly\FilamentBlog\Resources\Newsletters.
- Introduced new schemas and pages for managing newsletters, including CreateNewsletter, EditNewsletter, and ListNewsletters.
- Removed the deprecated NewsletterResource class and its associated methods to enhance clarity and maintainability.
- Updated Blog.php and other related files to reflect the new NewsletterResource path.
