--- src/Resources/NewsletterResource.php
@@ -1,92 +1,0 @@
-<?php
-
-namespace Firefly\FilamentBlog\Resources;
-
-use Filament\Schemas\Schema;
-use Filament\Forms\Components\TextInput;
-use Filament\Forms\Components\Toggle;
-use Filament\Tables\Columns\TextColumn;
-use Filament\Tables\Columns\ToggleColumn;
-use Filament\Actions\EditAction;
-use Filament\Actions\BulkActionGroup;
-use Filament\Actions\DeleteBulkAction;
-use Firefly\FilamentBlog\Resources\NewsletterResource\Pages\ListNewsletters;
-use Firefly\FilamentBlog\Resources\NewsletterResource\Pages\CreateNewsletter;
-use Firefly\FilamentBlog\Resources\NewsletterResource\Pages\EditNewsletter;
-use Filament\Forms;
-use Filament\Resources\Resource;
-use Filament\Tables;
-use Filament\Tables\Table;
-use Firefly\FilamentBlog\Models\NewsLetter;
-
-class NewsletterResource extends Resource
-{
-    protected static ?string $model = NewsLetter::class;
-
-    protected static string | \BackedEnum | null $navigationIcon = 'heroicon-o-newspaper';
-
-    protected static string | \UnitEnum | null $navigationGroup = 'Blog';
-
-    protected static ?int $navigationSort = 6;
-
-    public static function form(Schema $schema): Schema
-    {
-        return $schema
-            ->components([
-                TextInput::make('email')
-                    ->email()
-                    ->required()
-                    ->unique(ignoreRecord: true)
-                    ->maxLength(100),
-                Toggle::make('subscribed')
-                    ->default(true)
-                    ->required()->columnSpanFull(),
-            ])->columns(2);
-    }
-
-    public static function table(Table $table): Table
-    {
-        return $table
-            ->columns([
-                TextColumn::make('email')
-                    ->searchable(),
-                ToggleColumn::make('subscribed')
-                    ->label('Subscribed'),
-                TextColumn::make('created_at')
-                    ->dateTime()
-                    ->sortable()
-                    ->toggleable(isToggledHiddenByDefault: true),
-                TextColumn::make('updated_at')
-                    ->dateTime()
-                    ->sortable()
-                    ->toggleable(isToggledHiddenByDefault: true),
-            ])
-            ->filters([
-                //
-            ])
-            ->recordActions([
-                EditAction::make(),
-            ])
-            ->toolbarActions([
-                BulkActionGroup::make([
-                    DeleteBulkAction::make(),
-                ]),
-            ]);
-    }
-
-    public static function getRelations(): array
-    {
-        return [
-            //
-        ];
-    }
-
-    public static function getPages(): array
-    {
-        return [
-            'index' => ListNewsletters::route('/'),
-            'create' => CreateNewsletter::route('/create'),
-            'edit' => EditNewsletter::route('/{record}/edit'),
-        ];
-    }
-}
